+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Better error handling and test cleanup ([@zvonimirr](https://github.com/zvonimirr))
 - Set & list preferences API method ([@zvonimirr](https://github.com/zvonimirr))
 - Implement on-the-fly currency conversion ([@zvonimirr](https://github.com/zvonimirr))
+- Properties now return monthly revenue prediction ([@zvonimirr](https://github.com/zvonimirr))
 
 ### UI
 - Rent status is now displayed ([@kovaj024](https://github.com/kovaj024))